refactor(analytics): Split anomaly and prediction modules
Refactored analytics modules to follow single responsibility principle:

**Anomaly Module (302 lines → 4 files):**
- types.rs (51 lines): AnomalyDetectionAlgorithm, AnomalyScore
- detector.rs (254 lines): AnomalyDetector implementation
- realtime.rs (69 lines): RealtimeAnomalyDetector
- mod.rs (10 lines): Re-exports only

**Prediction Module (259 lines → 4 files):**
- types.rs (48 lines): TrendDirection, PredictionResult
- predictor.rs (170 lines): TimeSeriesPredictor implementation
- trend.rs (136 lines): TrendDetector implementation
- mod.rs (10 lines): Re-exports only

**Testing:**
- All 408 tests passing
- No clippy warnings
- Formatting check passed
- Public API unchanged

Closes #162
